feat: refactored buy key
